Add signal support
The meaurement values returned by a driver are referred to as Signals.
Typically a signal performs a simple transformation on the raw value. It
has a name and the value can be printed.
We add a base class that can be used to simple configure new signals
follow this scheme.